agent-api/requirements.txt
fastapi>=0.112.0
uvicorn[standard]>=0.30.0
python-multipart
pydantic>=2.7
python-dotenv
openai>=1.40.0
